moved Rectangle object to the domain package.
